Added water balloon materials
Hey guys! xD Sorry for kind of derping out with this one. I probably would have added the materials earlier, but I had to leave in the middle of all this to go to work.

The waterballoon asset should now be fully implemented in the repo.
(I also made some changes to the camera and light when I was looking at my model, so be warned.)